Fix crc includes

+++ Etape8/structure.h
@@ -12,5 +12,5 @@ struct RequeteBDEF {
     int NumeroTicket;
     int Heure;
     enum Action Action;
-    short CRC;
+    int CRC;
 };
